refactor: CancelReason application -> domain 계층으로 이동 및 @Enumerated 추가
