+++ src/main/java/com/goggles/orderservice/domain/model/CancelReason.java
@@ -1,4 +1,4 @@
-package com.goggles.orderservice.application.common;
+package com.goggles.orderservice.domain.model;
 
 import com.goggles.common.exception.BadRequestException;
